Add increments
