+++ data/alpha.txt
@@ -1,5 +1,3 @@
-T | 0 | go
-T | 0 | bye
-D | 0 | if | 2025-07-19 1900
-E | 0 | party | 2025-07-19 1900 | 2025-07-19 2100
-E | 0 | error | 2025-07-19 2100 | 2025-07-19 1900
+T | 0 | 1
+T | 0 | 2
+E | 0 | lunch | 2025-01-31 1306 | 2025-01-31 1330
